Rename products package into items
