+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/home/items/WooPosItemNavigationData.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.woopos.home.products
+package com.woocommerce.android.ui.woopos.home.items
 
 sealed class WooPosItemNavigationData(open val id: Long) {
     data class SimpleProductData(
