+++ WooCommerce/src/test/kotlin/com/woocommerce/android/ui/woopos/home/products/WooPosProductsDataSourceTest.kt
@@ -5,6 +5,7 @@ import com.woocommerce.android.ui.products.ProductTestUtils
 import com.woocommerce.android.ui.products.ProductType
 import com.woocommerce.android.ui.products.selector.ProductListHandler
 import com.woocommerce.android.ui.woopos.featureflags.IsNonSimpleProductTypesEnabled
+import com.woocommerce.android.ui.woopos.home.items.WooPosProductsDataSource
 import com.woocommerce.android.ui.woopos.util.WooPosCoroutineTestRule
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.first
